+++ src/main/resources/templates/user/changeinfo.html
@@ -18,7 +18,7 @@
                         <div class="bg-secondary rounded p-4 p-sm-5 my-4 mx-3">
                             <form>
                                 <div class="d-flex align-items-center justify-content-between mb-3">
-                                    <a href="index.html">
+                                    <a href="../project/index.html">
                                     <h3 class="text-primary"><i class="fa fa-user-edit me-2"></i>LostArkProject</h3>
                                     </a>
                                     <h3>회원정보</h3>
