+++ src/main/resources/templates/user/signup.html
@@ -13,7 +13,7 @@
       <div class="col-12 col-sm-8 col-md-6 col-lg-5 col-xl-4">
         <div class="bg-secondary rounded p-4 p-sm-5 my-4 mx-3">
           <div class="d-flex align-items-center justify-content-between mb-3">
-            <a href="index.html" class="">
+            <a href="../project/index.html" class="">
               <h3 class="text-primary"><i class="fa fa-user-edit me-2"></i>LostArkProject</h3>
             </a>
             <h3>회원가입</h3>
